// front-end/src/Models/NumidentHint.ts
export default interface NumidentHint {
  ark: string;
  batchID: string;
  birthdate: string;
  birthplace: string;
  birthstate: string;
  bir_state: string;
  deathdate: string;
  father_ark: string;
  father_givenname: string;
  father_pid: string;
  father_surname: string;
  givenname: string;
  mother_ark: string;
  mother_givenname: string;
  mother_pid: string;
  mother_surname: string;
  pid: string;
  pr_bir_date: string;
  pr_bir_place: string;
  pr_dea_date: string;
  pr_fthr_name_gn: string;
  pr_fthr_name_gn_unique: number;
  pr_fthr_name_surn: string;
  pr_fthr_name_surn_unique: number;
  pr_mthr_name_gn: string;
  pr_mthr_name_gn_unique: number;
  pr_mthr_name_surn: string;
  pr_mthr_name_surn_unique: number;
  pr_name_gn: string;
  pr_name_gn_unique: number;
  pr_name_surn: string;
  pr_name_surn_unique: number;
  results: string;
  score: number;
  surname: string;
  uploadTime: number;
  url: string;
}